[feature] move SDDDriver class to new file
